# public/api-spec.yaml
openapi: 3.0.3
info:
  title: DocGen Engine API
  description: |
    REST API for generating PDF business documents from templates and JSON data.
    
    ### Features
    - Generate PDF from HTML templates
    - Scan DOCX templates to detect fields
    - Generate PDF from DOCX with data injection
    - LLM-powered field mapping and validation
    - Watermark support
  version: 1.0.0
  contact:
    name: DocGen Support
    url: https://github.com/fikriaf/DocGen
servers:
  - url: https://docgen-production-503d.up.railway.app/api/v1
    description: Production Server
  - url: http://localhost:3000/api/v1
    description: Development Server
...
paths:
...
  /docx/scan:
    post:
      tags:
        - Scanning
      summary: Scan DOCX template for fields
      description: |
        Upload a DOCX file to automatically detect all variables/fields that need to be filled.
        
        Supports both explicit markers like `{{field_name}}` and can use AI to detect
        fields from plain text documents.
      operationId: scanDocxTemplate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: DOCX file to scan
      responses:
        '200':
          description: Fields detected successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  fields:
                    type: array
                    items:
                      type: string
                    example:
                      - invoice_no
                      - client_name
                      - total_amount
  /docx/generate:
    post:
      tags:
        - Generation
      summary: Generate PDF from DOCX template
      description: |
        Upload a DOCX template and inject JSON data to generate a PDF.
        
        - Injects payload JSON into DOCX variables
        - Converts DOCX to HTML
        - Renders HTML to PDF
        - Optionally applies watermark
      operationId: generateDocxTemplate
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - payload
              properties:
                file:
                  type: string
                  format: binary
                  description: DOCX template file
                payload:
                  type: object
                  description: JSON object containing field values to inject
                  example:
                    invoice_no: "INV-2026-001"
                    client_name: "John Doe"
                    date: "2026-03-13"
                options:
                  type: object
                  description: Optional JSON object for watermark and PDF settings
                  example:
                    watermark:
                      enabled: true
                      text: "DRAFT"
                      color: "#FF0000"
            encoding:
              payload:
                contentType: application/json
              options:
                contentType: application/json
      responses:
        '200':
          description: PDF generated successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
...